Move PeerPool into seperate file
